requirements.txt
streamlit
scikit-learn
pandas
numpy
plotly
joblib
biopython
